backend/requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langgraph
sqlalchemy
python-dotenv
pydantic